requirements.txt
streamlit
pandas
numpy
plotly
reportlab
python-docx
